+++ edx_toggles/toggles/internal/waffle/legacy.py
@@ -1,36 +1,29 @@
 """
-Waffle classes in the context of edx-platform and other IDAs.
+This module contains legacy code for backward compatibility. Waffle flag and switch objects previously required the
+creation of namespace objects. The namespace features were all moved to the WaffleSwitch/Flag classes.
 
-Includes namespacing and caching for waffle flags.
+To upgrade your code, use the following guidelines. Where previously you had::
 
-Usage:
+    from edx_toggles.toggles import WaffleSwitch, WaffleSwitchNamespace
+    SOME_NAMESPACE = WaffleSwitchNamespace("some_namespace")
+    SOME_SWITCH = WaffleSwitch(SOME_NAMESPACE, "some_switch", module_name=__name__)
 
-For Waffle Flags, first set up the namespace, and then create flags using the
-namespace.  For example::
+You should now write::
 
-   WAFFLE_FLAG_NAMESPACE = WaffleFlagNamespace(name='my_namespace')
-   SOME_FLAG = WaffleFlag(WAFFLE_FLAG_NAMESPACE, 'some_feature', __name__)
+    from edx_toggles.toggles.__future__ import WaffleSwitch
+    SOME_SWITCH = WaffleSwitch("some_namespace.some_switch", module_name=__name__)
 
-You can check theis flag in code using the following::
+And similarly for waffle flags, replace::
 
-    SOME_FLAG.is_enabled()
+    from edx_toggles.toggles import WaffleFlag, WaffleFlagNamespace
+    SOME_NAMESPACE = WaffleFlagNamespace("some_namespace", log_prefix="some_namespace")
+    SOME_FLAG = WaffleFlag(SOME_NAMESPACE, "some_flag", module_name=__name__)
 
-To test these WaffleFlags, see testutils.py.
+by::
 
-In the above examples, you will use Django Admin "waffle" section to configure
-for a flag named: my_namespace.some_course_feature
-
-For Waffle Switches, follow the example above for WaffleFlags, but instead use the WaffleSwitchNamespace and
-WaffleSwitch classes.
-
-For long-lived flags, you may want to change the default for devstack, sandboxes,
-or new Open edX releases. For help with this, see:
-openedx/core/djangoapps/waffle_utils/docs/decisions/0001-refactor-waffle-flag-default.rst
-
-Also see ``WAFFLE_FLAG_CUSTOM_ATTRIBUTES`` and docstring for _set_waffle_flag_attribute
-for temporarily instrumenting/monitoring waffle flag usage.
+    from edx_toggles.toggles.__future__ import WaffleFlag
+    SOME_FLAG = WaffleFlag("some_namespace.some_flag", module_name=__name__, log_prefix="some_namespace")
 """
-import warnings
 from abc import ABC
 
 from edx_django_utils.monitoring import set_custom_attribute
@@ -59,30 +52,25 @@ class BaseNamespace(ABC):
                 (e.g. "Grades: "). Defaults to ''.
 
         """
-        warnings.warn(
-            (
-                "{} is deprecated. Please use non-namespaced edx_toggles.toggles.WaffleFlag/WaffleSwitch"
-                " classes instead."
-            ).format(self.__class__.__name__),
-            DeprecationWarning,
-            stacklevel=2,
+        set_custom_attribute(
+            self.__class__.__module__,
+            "{}[{}]".format(self.__class__.__name__, name),
         )
-        set_custom_attribute("deprecated_edx_toggles_waffle", self.__class__.__name__)
         assert name, "The name is required."
         self.name = name
         self.log_prefix = log_prefix if log_prefix else ""
 
-    def _namespaced_name(self, setting_name):
+    def _namespaced_name(self, toggle_name):
         """
         Returns the namespaced name of the waffle switch/flag.
 
         For example, the namespaced name of a waffle switch/flag would be:
-            my_namespace.my_setting_name
+            my_namespace.my_toggle_name
 
         Arguments:
-            setting_name (String): The name of the flag or switch.
+            toggle_name (String): The name of the flag or switch.
         """
-        return "{}.{}".format(self.name, setting_name)
+        return "{}.{}".format(self.name, toggle_name)
 
 
 class WaffleSwitchNamespace(BaseNamespace):
@@ -92,51 +80,54 @@ class WaffleSwitchNamespace(BaseNamespace):
 
     def is_enabled(self, switch_name):
         """
-        Legacy method preserved for backward compatibility.
+        Returns whether or not the switch is enabled.
         """
-        return NewWaffleSwitch(self._namespaced_name(switch_name)).is_enabled()
-
-    def get_request_cache(self, namespaced_switch_name, default=None):
-        return NewWaffleSwitch(namespaced_switch_name).get_request_cache(
-            default=default
-        )
-
-    def get_request_cache_with_short_name(self, switch_name, default=None):
-        return self.get_request_cache(
-            self._namespaced_name(switch_name), default=default
-        )
-
-    def set_request_cache(self, namespaced_switch_name, value):
-        NewWaffleSwitch(namespaced_switch_name).set_request_cache(value)
+        return NewWaffleSwitch(self._namespaced_name(switch_name), __name__).is_enabled()
 
     def set_request_cache_with_short_name(self, switch_name, value):
-        self.set_request_cache(self._namespaced_name(switch_name), value)
+        """
+        Explicitly set request cache value for the (non-namespaced) switch name. You should avoid using this method as
+        much as possible as it may have unexpected side-effects.
+        """
+        namespaced_name = self._namespaced_name(switch_name)
+        # pylint: disable=protected-access
+        NewWaffleSwitch(namespaced_name, __name__)._cached_switches[namespaced_name] = value
 
 
 class WaffleSwitch(NewWaffleSwitch):
     """
-    Legacy namespaced waffle switch class.
+    Represents a single waffle switch, enhanced with request level caching.
     """
 
     def __init__(self, waffle_namespace, switch_name, module_name=None):
-        warnings.warn(
-            (
-                "{} is deprecated. Please use non-namespaced edx_toggles.toggles.WaffleSwitch instead."
-            ).format(self.__class__.__name__),
-            DeprecationWarning,
-            stacklevel=2,
-        )
-        set_custom_attribute("deprecated_edx_toggles_waffle", "WaffleSwitch")
         if not isinstance(waffle_namespace, str):
             waffle_namespace = waffle_namespace.name
+        set_custom_attribute("deprecated_edx_toggles_waffle", "WaffleSwitch")
 
-        # Non-namespaced flag_name attribute preserved for backward compatibility
-        self.switch_name = switch_name
+        self._switch_name = switch_name
         name = "{}.{}".format(waffle_namespace, switch_name)
         super().__init__(name, module_name=module_name)
 
     @property
+    def switch_name(self):
+        """
+        Non-namespaced switch_name attribute preserved for backward compatibility.
+        """
+        set_custom_attribute(
+            "deprecated_waffle_legacy_method",
+            "WaffleSwitch[{}].switch_name".format(self.name),
+        )
+        return self._switch_name
+
+    @property
     def namespaced_switch_name(self):
+        """
+        This is now equivalent to the switch name.
+        """
+        set_custom_attribute(
+            "deprecated_waffle_legacy_method",
+            "WaffleSwitch[{}].namespaced_switch_name".format(self.name),
+        )
         return self.name
 
 
@@ -149,7 +140,31 @@ class WaffleFlagNamespace(BaseNamespace):
         """
         Returns and caches whether the provided flag is active.
         """
-        return WaffleFlag(self, flag_name).is_enabled()
+        return NewWaffleFlag(
+            self._namespaced_name(flag_name), module_name=__name__
+        ).is_enabled()
+
+    def _monitor_value(self, flag_name, value):
+        """
+        Monitoring method preserved for backward compatibility. You should use `WaffleFlag.set_monitor_value` instead.
+        """
+        set_custom_attribute(
+            "deprecated_edx_toggles_waffle", "WaffleFlagNamespace._monitor_value"
+        )
+        return NewWaffleFlag(
+            self._namespaced_name(flag_name), module_name=__name__
+        ).set_monitor_value(value)
+
+    @property
+    def _cached_flags(self):
+        """
+        Legacy property used by CourseWaffleFlag.
+        """
+        set_custom_attribute(
+            "deprecated_waffle_legacy_method",
+            "WaffleFlagNamespace[{}]._cached_flags".format(self.name),
+        )
+        return NewWaffleFlag.cached_flags()
 
 
 class WaffleFlag(NewWaffleFlag):
@@ -158,13 +173,6 @@ class WaffleFlag(NewWaffleFlag):
     """
 
     def __init__(self, waffle_namespace, flag_name, module_name=None):
-        warnings.warn(
-            (
-                "{} is deprecated. Please use non-namespaced edx_toggles.toggles.WaffleFlag instead."
-            ).format(self.__class__.__name__),
-            DeprecationWarning,
-            stacklevel=2,
-        )
         set_custom_attribute("deprecated_edx_toggles_waffle", "WaffleFlag")
         log_prefix = ""
         if not isinstance(waffle_namespace, str):
@@ -172,13 +180,39 @@ class WaffleFlag(NewWaffleFlag):
             waffle_namespace = waffle_namespace.name
 
         # Non-namespaced flag_name attribute preserved for backward compatibility
-        self.flag_name = flag_name
+        self._flag_name = flag_name
         name = "{}.{}".format(waffle_namespace, flag_name)
         super().__init__(name, module_name=module_name, log_prefix=log_prefix)
+
+    @property
+    def flag_name(self):
+        """
+        Non-namespaced flag_name attribute preserved for backward compatibility.
+        """
+        set_custom_attribute(
+            "deprecated_waffle_legacy_method",
+            "WaffleFlag[{}].flag_name".format(self.name),
+        )
+        return self._flag_name
 
     @property
     def namespaced_flag_name(self):
         """
         Preserved for backward compatibility.
         """
+        set_custom_attribute(
+            "deprecated_waffle_legacy_method",
+            "WaffleFlag[{}].namespaced_flag_name".format(self.name),
+        )
         return self.name
+
+    @property
+    def waffle_namespace(self):
+        """
+        Preserved for backward compatibility.
+        """
+        set_custom_attribute(
+            "deprecated_waffle_legacy_method",
+            "WaffleFlag[{}].waffle_namespace".format(self.name),
+        )
+        return WaffleFlagNamespace(self.name.split(".")[0], log_prefix=self.log_prefix)