+++ edx_toggles/toggles/internal/waffle/switch.py
@@ -11,32 +11,20 @@ from .cache import _get_waffle_request_cache
 
 class WaffleSwitch(BaseWaffle):
     """
-    Represents a single waffle switch, using both a global and a request cache.
+    Represents a single waffle switch, enhanced with request-level caching.
     """
 
     _class_instances = WeakSet()
 
     def is_enabled(self):
         """
-        Legacy method preserved for backward compatibility.
+        Returns whether or not the switch is enabled.
         """
-        value = self.get_request_cache()
+        value = self._cached_switches.get(self.name)
         if value is None:
             value = switch_is_active(self.name)
-            self.set_request_cache(value)
+            self._cached_switches[self.name] = value
         return value
-
-    def get_request_cache(self, default=None):
-        """
-        API for accessing the request cache. In general, users should avoid accessing the namespace cache.
-        """
-        return self._cached_switches.get(self.name, default)
-
-    def set_request_cache(self, value):
-        """
-        Manually set the request cache value. Beware! There be dragons.
-        """
-        self._cached_switches[self.name] = value
 
     @property
     def _cached_switches(self):
@@ -44,3 +32,14 @@ class WaffleSwitch(BaseWaffle):
         Return a dictionary of all namespaced switches in the request cache.
         """
         return _get_waffle_request_cache().setdefault("switches", {})
+
+
+class NonNamespacedWaffleSwitch(WaffleSwitch):
+    """
+    Same as the WaffleSwitch class, but does not require that the instance name be namespaced. This class is useful for
+    migrating existing Switch objects; new instances should always be namespaced.
+    """
+
+    @classmethod
+    def validate_name(cls, name):
+        pass